Update ping.py
Added a try-except block to validate the host argument using socket.getaddrinfo(). If the argument is not a valid hostname or IP address, the script exits with an error message.

Added support for continuous pinging with the -c or --count option. If the option is set to a value greater than 1, the script sends the specified number of packets at the specified interval using the -i or --interval option. The script only exits if an error occurs or the user interrupts the script using Ctrl+C.

Added support for IPv6 addresses with the --ipv6 option. If the option is set, the script uses IPv6 instead of IPv4.

Added support for specifying the number of data bytes to be sent with the -s or --size option.

Added support for quiet mode with the -q or --quiet option. In quiet mode, only the summary output is displayed, and nothing else is printed to the console.

Added support for TCP and UDP pinging modes with the -m or --mode option. The default mode is icmp, but you can also use tcp or udp modes to test network connectivity with different protocols.

Added support for outputting ping results to a CSV or JSON file with the -f or --file option. If the option is set, the script saves the results to a file with the specified name and format. If the file already exists, the script appends the results to the end of the file.

Improved the summary output to include packet loss percentage and standard deviation, in addition to the minimum, maximum, and average round-trip times.